staging: wilc1000: refactor code to remove 'mac_status' from 'wilc_mac_cfg' struct
Refactor the code by removing use of 'mac_status' from 'wilc_mac_cfg'
and only have the string type configruation values in that struct.
Now fetch the value 'WID_STATUS' configuration from 'g_cfg_byte' array.

+++ wilc/wilc_wfi_netdevice.h
@@ -177,7 +177,7 @@ struct wilc_vif {
 struct wilc {
 	const struct wilc_hif_func *hif_func;
 	int io_type;
-	int mac_status;
+	s8 mac_status;
 	int gpio_irq;
 	int gpio_reset;
 	int gpio_chip_en;
